Create extensions `backupStateFlow()` and `recoveryStateFlow`.
